requirements.txt
# Backend dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
httpx==0.25.2
python-multipart==0.0.6
firebase-admin==6.2.0
python-dotenv==1.0.0
